Reimplement test performers to use throw
